+++ tests/src/Functional/DataService/Media/MediaQueryTest.php
@@ -27,7 +27,7 @@ class MediaQueryTest extends FunctionalTestBase
         $filter->getRange()
             ->setStartIndex(1)
             ->setEndIndex(2);
-        $results = $dof->select($filter, $this->account);
+        $results = $dof->select($filter);
 
         foreach ($results as $index => $result) {
             $this->assertInstanceOf(UriInterface::class, $result->getId());
@@ -65,12 +65,12 @@ class MediaQueryTest extends FunctionalTestBase
             ->setStartIndex(1)
             ->setEndIndex(2);
         /** @var ObjectList $list */
-        $list = $dof->selectRequest($filter, $this->account)->wait();
+        $list = $dof->selectRequest($filter)->wait();
 
         // Determine the end of the list of ranges.
         $ranges = Range::nextRanges($list);
         $filter->setRange(end($ranges));
-        $results = $dof->select($filter, $this->account);
+        $results = $dof->select($filter);
 
         foreach ($results as $index => $result) {
             $this->assertInstanceOf(UriInterface::class, $result->getId());